# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Sightquest server",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/styles/me/avatar": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "style"
                ],
                "summary": "Set avatar style",
                "parameters": [
                    {
                        "description": "Style ID",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/style.setAvatarRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.User": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "styles": {
                    "$ref": "#/definitions/domain.UserStyles"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "domain.UserStyles": {
            "type": "object",
            "properties": {
                "avatarId": {
                    "type": "string"
                }
            }
        },
        "style.setAvatarRequest": {
            "type": "object",
            "properties": {
                "styleId": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Token",
            "in": "header"
        }
    }
}